Move in-memory Source to its own file
